# cdc_erp_updated/cdc_erpp1/cdc_erpp/Document-Type-Designs/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CDC ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all case files
      parameters:
        - name: childId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseFile"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseFile"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case file by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseFile"
    put:
      operationId: updateCase
      tags: [cases]
      summary: Update a case file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseFile"

  # Health Assessments
...
  /court-cases:
    get:
      operationId: listCourtCases
      tags: [court-cases]
      summary: List court cases
      parameters:
        - name: childId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of court cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourtCase"
    post:
      operationId: createCourtCase
      tags: [court-cases]
      summary: Create a court case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourtCaseBody"
      responses:
        "201":
          description: Created court case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtCase"
  /court-cases/{id}:
    get:
      operationId: getCourtCase
      tags: [court-cases]
      summary: Get a court case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Court case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtCase"
    put:
      operationId: updateCourtCase
      tags: [court-cases]
      summary: Update a court case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourtCaseBody"
      responses:
        "200":
          description: Updated court case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtCase"
  /court-cases/bulk-import:
    post:
      operationId: bulkImportCourtCases
      tags: [court-cases]
      summary: Bulk import court cases from CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file containing court case data
      responses:
        "200":
          description: Bulk import results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: integer
                    description: Number of successfully imported records
                  errors:
                    type: array
                    items:
                      type: string
                    description: List of error messages for failed imports
                  total:
                    type: integer
                    description: Total number of records processed

  # Police Acquisitions
...
components:
  schemas:
...
    CaseFile:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: string
        childId:
          type: integer
        childName:
          type: string
        caseOpeningDate:
          type: string
          format: date
        assignedCaseWorker:
          type: string
        riskLevel:
          type: string
          enum: [Low, Medium, High]
        caseStatus:
          type: string
          enum: [Open, Active, Review, Closed]
        caseSummary:
          type: string
        investigationNotes:
          type: string
        recommendation:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - childId
        - caseStatus
        - createdAt
        - updatedAt
    CreateCaseBody:
      type: object
      properties:
        childId:
          type: integer
        caseOpeningDate:
          type: string
          format: date
        assignedCaseWorker:
          type: string
        riskLevel:
          type: string
          enum: [Low, Medium, High]
        caseStatus:
          type: string
          enum: [Open, Active, Review, Closed]
        caseSummary:
          type: string
        investigationNotes:
          type: string
        recommendation:
          type: string
      required:
        - childId
        - caseStatus
...
    CourtCase:
      type: object
      properties:
        id:
          type: integer
        courtCaseId:
          type: string
        childId:
          type: integer
        childName:
          type: string
        courtName:
          type: string
        policeStationName:
          type: string
        grNumber:
          type: string
        caseNo:
          type: string
        legalSection:
          type: string
        legalAidType:
          type: string
        hearingDate:
          type: string
          format: date
        lastHearingDate:
          type: string
          format: date
        lawyerName:
          type: string
        childCaseType:
          type: string
        previousCaseInvolvement:
          type: boolean
        outcome:
          type: string
        nextHearingDate:
          type: string
          format: date
        firNumber:
          type: string
        firDate:
          type: string
          format: date
        currentCaseStatus:
          type: string
        courtAttendanceDetails:
          type: string
        courtAttendanceDates:
          type: string
        guardianCommunication:
          type: string
        educationTraining:
          type: string
        centerFacilities:
          type: string
        caseComments:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - courtCaseId
        - childId
        - courtName
        - caseNo
        - createdAt
    CreateCourtCaseBody:
      type: object
      properties:
        childId:
          type: integer
        courtName:
          type: string
        policeStationName:
          type: string
        grNumber:
          type: string
        caseNo:
          type: string
        legalSection:
          type: string
        legalAidType:
          type: string
        hearingDate:
          type: string
          format: date
        lastHearingDate:
          type: string
          format: date
        lawyerName:
          type: string
        childCaseType:
          type: string
        previousCaseInvolvement:
          type: boolean
        outcome:
          type: string
        nextHearingDate:
          type: string
          format: date
        firNumber:
          type: string
        firDate:
          type: string
          format: date
        currentCaseStatus:
          type: string
        courtAttendanceDetails:
          type: string
        courtAttendanceDates:
          type: string
        guardianCommunication:
          type: string
        educationTraining:
          type: string
        centerFacilities:
          type: string
        caseComments:
          type: string
      required:
        - childId
        - courtName
        - caseNo
...